Initialize agents to simulate
Provide agents with a random position and a single agent chosen as the
agent to be 'It'. This is needed as an initial situation to start the
Tag simulation from.
It is easier to build this out in a different module to keep
CLI/settings logic separate from the simulation logic.