Agents walk around randomly within the bounds
Make the agents take a random action each simulation step, this will
move them around the bounded area. For each time an agent tries to move
out of the bounds, let it remain at the edge of the boundary.

The lower bounds are always 0, the upper bounds are given by the
settings provided.